Row count and column counts are integer types.

+++ Matrix.h
@@ -1,11 +1,11 @@
 #pragma once
 typedef struct Matrix
 {
-	short rowCount;
-	short columnCount;
+	int rowCount;
+	int columnCount;
 	float* data;
 } Matrix;
 
-// Parametre olarak verilen iki matrisi toplar ve döndürür.
+// Parametre olarak verilen iki matrisi toplar ve sonucu döndürür.
 // Eðer matrislerin boyutu birbirine eþit deðilse NULL (0) döndürür.
 Matrix* matrix_addition(Matrix*, Matrix*);
